Move Trigger interface to domain layer
