+++ src/triggers/ReplyTrigger.ts
@@ -7,7 +7,7 @@ import type {
   Trigger,
   TriggerContext,
   TriggerResult,
-} from './Trigger.interface';
+} from '../domain/triggers/Trigger.interface';
 
 export class ReplyTrigger implements Trigger {
   private readonly logger: Logger;
